refactor: 날짜 파싱 유틸 정리 및 steps 타입 변경

+++ src/main/java/com/kb/healthcare/myohui/domain/dto/HealthDataRequest.java
@@ -40,7 +40,7 @@ public class HealthDataRequest {
         private Period period;
         private Metric distance;
         private Metric calories;
-        private int steps;
+        private float steps;
 
         @Getter
         public static class Period {
